requirements.txt
fastapi
uvicorn
motor
python-multipart
# for authentication
python-jose
passlib[bcrypt]